Create notes.c
Next open up notes.c. In this file is a program (soon to be called notes) that not only prints the frequencies (in Hz) of all of the notes in an octave, it also outputs a WAV file (an audio file) via which you can hear those same notes. By default, it does so for the fourth octave, but if you pass it a command-line argument (a number between 0 and 8, inclusive), you can see and hear the frequencies of any octave’s notes.

Read through the comments and code in notes.c and try to understand most, if not all, of its lines. Some might look unfamiliar. For instance, by convention, it uses a function called fprintf to print error messages to stderr (aka standard error) rather than printf, which, it turns out, prints to something called stdout (aka standard output). By default, messages printed to stdout and stderr both appear on the user’s screens. But it’s possible to separate them when running a program so that users can distinguish error messages from non-error messages. But more on that perhaps another time!

Notice, too, how main returns 1 in cases of error. That, too, is a convention. To date, we’ve not returned any values from main. But, recall that, all this time, main has had a return type, specifically int. It turns out, when main is done executing, it returns 0 by default, which, by convention, signifies success. If something goes wrong in a program, though, it’s convention to return some value other than 0 (e.g., 1). That value is called an "exit code" and can be used to distinguish one type of error from another. In fact, if you’ve ever seen a cryptic error code on your Mac’s or PC’s screen, it might very well have been the value returned by some (buggy) program’s main function.

Notice too how this program uses a function called sprintf which doesn’t actually print to the screen but instead stores its output in a string (hence the s in sprintf). We’re using it in order to create a string from two placeholders, %s and %i. Notice how we allocate space for a (short) string by declaring an array for 4 chars. We then use sprintf to store a NOTES[i] (a string, ergo the %s) in that memory followed by octave (an int, ergo the %i). That way, we can take values like "A" and 4 and, effectively, concatenate them (i.e., append the latter to the former) in order to create a new string, the value of which is, for instance, A4.

Along the way in this program do we call some (presumably) unfamilar functions called song_open, frequency, note_write, and song_close. It turns out those functions are implemented in other files in this problem’s distribution. Keep an eye out for them!