Create synthesize.c
In this file is a program (soon to be called synthesize) that synthesizes (i.e., generates) a song from a sequence of notes. Notice how it gets those notes from a user one at a time using get_string. It first checks, though, whether the user’s input is a rest, as would happen if the user simply hits Enter. Else it proceeds to "tokenize" the user’s input into two tokens: a note, which can be found to the left of the @ in the user’s input, and a fraction, which can be found to the right of the @ in the user’s input. The program uses a function called strtok to facilitate such. It then writes that note (or rest) to a file.